Refactor project structure into standard Python package structure
