+++ cimgen/build.py
@@ -1,0 +1,28 @@
+import argparse
+import importlib
+import os
+
+from cimgen import cimgen
+
+
+def build():
+    parser = argparse.ArgumentParser(description="Generate some CIM classes.")
+    parser.add_argument("--outdir", type=str, help="The output directory", required=True)
+    parser.add_argument("--schemadir", type=str, help="The schema directory", required=True)
+    parser.add_argument("--langdir", type=str, help="The langpack directory", required=True)
+    parser.add_argument(
+        "--cgmes_version",
+        type=str,
+        choices=["cgmes_v2_4_13", "cgmes_v2_4_15", "cgmes_v3_0_0"],
+        default="cgmes_v2_4_15",
+        help="CGMES Version",
+    )
+    args = parser.parse_args()
+
+    langPack = importlib.import_module(f"cimgen.languages.{args.langdir}.lang_pack")
+    schema_path = os.path.join(os.getcwd(), args.schemadir)
+    cimgen.cim_generate(schema_path, args.outdir, args.cgmes_version, langPack)
+
+
+if __name__ == "__main__":
+    build()
