+++ cimgen/languages/cpp/templates/cpp_enum_header_template.mustache
