+++ cimgen/languages/cpp/templates/cpp_enum_object_template.mustache
