+++ cimgen/languages/cpp/templates/cpp_float_header_template.mustache
