+++ cimgen/languages/cpp/templates/cpp_float_object_template.mustache
