+++ cimgen/languages/cpp/templates/cpp_header_template.mustache
