+++ cimgen/languages/cpp/templates/cpp_object_template.mustache
