+++ cimgen/languages/java/AttributeInterface.java
