+++ cimgen/languages/java/BaseClass.java
