+++ cimgen/languages/java/BaseClassBuilder.java
