+++ cimgen/languages/java/BaseClassInterface.java
