+++ cimgen/languages/java/templates/java_class.mustache
