+++ cimgen/languages/java/templates/java_enum.mustache
