+++ cimgen/languages/java/templates/java_float.mustache
