+++ cimgen/languages/javascript/templates/handlebars_baseclass_template.mustache
