+++ cimgen/languages/javascript/templates/handlebars_template.mustache
