+++ cimgen/languages/modernpython/templates/cimpy_class_template.mustache
