+++ cimgen/languages/modernpython/utils/base.py
