+++ cimgen/languages/modernpython/utils/config.py
@@ -7,5 +7,5 @@ from pydantic import ConfigDict
 cgmes_resource_config: ConfigDict = {
     # By default, with pydantic extra arguments given to a dataclass are silently ignored.
     # This matches the default behaviour by failing noisily.
-    "extra" : "forbid"
+    "extra": "forbid"
 }
