+++ cimgen/languages/modernpython/utils/constants.py
