+++ cimgen/languages/modernpython/utils/profile.py
