+++ cimgen/languages/python/templates/cimpy_class_template.mustache
