+++ pyproject.toml
@@ -1,20 +1,48 @@
-[tool.poetry]
+[project]
 name = "cimgen"
 version = "0.1.0"
 description = "Code generation from CIM data model for several programming languages"
-authors = []
-license = "Apache License 2.0"
 readme = "README.md"
+keywords = ["cim", "cgmes", "code-generation"]
 
-[tool.poetry.dependencies]
-python = "^3.10"
-poetry = ">=1.5.0" # key 'priority' in source
-xmltodict = "^0.13.0"
-chevron = "^0.14.0"
-cmake = "^3.27.0"
-pydantic = "<2"
-beautifulsoup4 = "^4.12.2"
+dependencies = [
+    "xmltodict >= 0.13.0, < 1",
+    "chevron >= 0.14.0, < 1",
+    "cmake >= 3.27.0, < 4",
+    "pydantic < 2",
+    "beautifulsoup4 >= 4.12.2, < 5"
+]
+
+requires-python = ">=3.11"
+
+classifiers = [
+  "Development Status :: 5 - Production/Stable",
+
+  "License :: OSI Approved :: Apache Software License",
+
+  "Programming Language :: Python :: 3.8",
+  "Programming Language :: Python :: 3.9",
+  "Programming Language :: Python :: 3.10",
+  "Programming Language :: Python :: 3.11"
+]
+
+[project.urls]
+Homepage = "https://github.com/sogno-platform/cimgen"
+Repository = "https://github.com/sogno-platform/cimgen"
+Issues = "https://github.com/sogno-platform/cimgen/issues"
+
+[project.scripts]
+cimgen = "cimgen.build:build"
 
 [build-system]
-requires = ["poetry-core"]
-build-backend = "poetry.core.masonry.api"
+requires = ["setuptools >= 61.0"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools]
+include-package-data = true
+
+[tool.setuptools.packages.find]
+include = ["cimgen*"]
+
+[tool.black]
+line-length = 120
